fixes jasp already installed (#1)
Updates jasp to check to see if the program is already installed before installing/updating. Though chocolatey does this already, the JASP installer will throw a notice indicating it has already installed and return a `2` for the exit code. The user has to interact with the notice to resume the chocolatey install. Current version is now installed appropriately or a message is given through chocolatey that there is nothing to install.